Version 1.0
This is the base code for the next implementations. 
there are classes :
- Allien class 
- Allien Fleet class
- Ship class
- Game class
- Bullet class
The space inavader is the main file that need to be lunch.
This game includes features such that we clan play the game put it is not the optimal game.
You can moove freely from right to left, you can shoot bullet to the allien that are descending.